refactor: name 수정 및 viewtype 경로 수정
