+++ src/features/feed/ui/header/MyFeedCatagory.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { View } from 'react-native';
-import Chip from '@shared/ui/Chip';
+import CategoryChipButton from '@shared/ui/CategoryChipButton';
 
 const MyFeedCatagory = () => {
   // FIXME: api로 전환
@@ -12,13 +12,17 @@ const MyFeedCatagory = () => {
     <View className="mt-1 h-[50px] w-full flex-row items-center justify-between px-1">
       <View className="w-[80%] flex-row items-center">
         {/* 전체 버튼 */}
-        <Chip label={'전체'} selected={selected === '전체'} onPress={() => setSelected('전체')} />
+        <CategoryChipButton
+          label={'전체'}
+          selected={selected === '전체'}
+          onPress={() => setSelected('전체')}
+        />
 
         <View className="flex-1 flex-row items-center">
           <View className="flex-1">
             <View className="flex-row items-center gap-2">
               {catagory.map((category, index) => (
-                <Chip
+                <CategoryChipButton
                   key={index}
                   label={category}
                   selected={selected === category}
