+++ src/pages/recipe/ui/MyRecipeCatagory.tsx
@@ -3,9 +3,9 @@ import { View, ScrollView } from 'react-native';
 
 import PlusIcon from '@/assets/img/plus.svg';
 import { useBottomSheetModal } from '@pages/recipe/model/useBottomSheetModal';
+import ViewTypeSwitcher from '@features/recipe/ui/ViewTypeSwitcher';
 import { useViewTypeStore } from '@shared/store/useViewTypeStore';
-import Chip from '@shared/ui/Chip';
-import ViewTypeSwitcher from '@shared/ui/viewType/ViewTypeSwitcher';
+import CategoryChipButton from '@shared/ui/CategoryChipButton';
 
 import MyRecipeCatagoryBottomSheet from './MyRecipeCatagoryBottomSheet';
 
@@ -22,7 +22,11 @@ const MyRecipeCatagory = () => {
       {/* 카테고리 UI */}
       <View className="mt-1 h-[50px] w-full flex-row items-center justify-between px-1">
         <View className="w-[80%] flex-row items-center">
-          <Chip label={'전체'} selected={selected === '전체'} onPress={() => setSelected('전체')} />
+          <CategoryChipButton
+            label={'전체'}
+            selected={selected === '전체'}
+            onPress={() => setSelected('전체')}
+          />
 
           <ScrollView
             horizontal
@@ -32,7 +36,7 @@ const MyRecipeCatagory = () => {
           >
             <View className="flex-row items-center gap-2">
               {catagory.map((category, index) => (
-                <Chip
+                <CategoryChipButton
                   key={index}
                   label={category}
                   selected={selected === category}
