+++ src/shared/types/view.ts
@@ -1,0 +1,1 @@
+export type ViewType = 'article' | 'feed' | 'image';
